Change variable type from int to double

+++ tp2/2_nested_functions/2_1_main.cpp
@@ -4,8 +4,8 @@ using namespace std;
 
 int main()
 {
-    int nbGrades;
-    int moyenne;
+    double nbGrades;
+    double moyenne;
 
     cout << "Entrez le nombre de notes : " <<endl;
     cin >> nbGrades;
